Feature: Dashboard page with user profile UI
