+++ app/dashboard/page.tsx
@@ -1,0 +1,26 @@
+import AuthButton from "@/components/AuthButton";
+import Header from "@/components/Header";
+import { createClient } from "@/utils/supabase/server";
+import { redirect } from "next/navigation";
+
+export default async function ProtectedPage() {
+    const supabase = createClient();
+
+    const {
+        data: { user },
+    } = await supabase.auth.getUser();
+
+    if (!user) {
+        return redirect("/login");
+    }
+
+    return (
+        <div className="flex-1 w-full flex flex-col gap-20 items-center">
+            <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
+                <div className="w-full max-w-4xl flex justify-between items-between p-3 text-sm">
+                    <AuthButton />
+                </div>
+            </nav>
+        </div>
+    );
+}
